+++ .github/workflows/release-to-changelog.yml
@@ -8,7 +8,7 @@ jobs:
   update-changelog:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
         with:
           ref: main
