+++ .github/workflows/size-limit.yml
@@ -14,7 +14,7 @@ jobs:
     env:
       CI_JOB_NUMBER: 1
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - uses: andresz1/size-limit-action@v1
         with:
           build_script: build:size
