+++ e2e/src/helper.ts
@@ -59,10 +59,6 @@ export async function activateAndWait(docUri: Uri): Promise<void> {
 }
 
 function onDidChangeTextDocument(e: TextDocumentChangeEvent): void {
-  if (e.document.uri.path.includes('Preview')) {
-    console.log(e.document.uri);
-  }
-
   if (e.document.uri.path === 'Preview' && previewPromiseResolve) {
     if (
       // When we switch to a new document, the preview content is set to '' we skip this such events here
